+++ PlataformaDenuncias/src/main/java/com/ucan/plataformadenuncias/repositories/ContaPerfilRepository.java
@@ -4,11 +4,6 @@ import com.ucan.plataformadenuncias.entities.ContaPerfil;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-/**
- *
- * @author cristiano
- */
 @Repository
-public interface ContaPerfilRepository extends JpaRepository<ContaPerfil, Long>{
-    
+public interface ContaPerfilRepository extends JpaRepository<ContaPerfil, Integer> {
 }
