+++ PlataformaDenuncias/src/main/java/com/ucan/plataformadenuncias/services/HistoricoActividadeService.java
@@ -25,11 +25,11 @@ public class HistoricoActividadeService {
         return repository.findAll();
     }
 
-    public Optional<HistoricoActividade> buscarPorId(Long id) {
+    public Optional<HistoricoActividade> buscarPorId(Integer id) {
         return repository.findById(id);
     }
 
-    public void remover(Long id) {
+    public void remover(Integer id) {
         repository.deleteById(id);
     }
 }
